+++ demo/lib/demo_web/router.ex
@@ -26,6 +26,7 @@ defmodule DemoWeb.Router do
       live "/forms", Live.FormsLive, :index
       live "/inputs", Live.InputsLive, :index
       live "/validations", Live.ValidationsLive, :index
+      live "/checkbox-lists", Live.CheckboxListsLive, :index
       live "/loaders", Live.LoadersLive, :index
       live "/lists", Live.ListsLive, :index
       live "/timeline", Live.TimelineLive, :index